+++ stacker/commands/stacker/build.py
@@ -37,9 +37,14 @@ class Build(BaseCommand):
                                  "specified more than once. If not specified "
                                  "then stacker will work on all stacks in the "
                                  "config file.")
-        parser.add_argument('-t', '--tail', action='store_true',
-                            help='Tail the CloudFormation logs while working'
-                                 'with stacks')
+        parser.add_argument("-t", "--tail", action="store_true",
+                            help="Tail the CloudFormation logs while working"
+                                 "with stacks")
+        parser.add_argument("--skip-hook", action="append", default=[],
+                            choices=["pre", "post"],
+                            help="Which hooks to skip during build")
+        parser.add_argument("--run-hook", choices=["pre", "post"],
+                            help="Run a hook without the rest of the build")
 
     def run(self, options, **kwargs):
         super(Build, self).run(options, **kwargs)
@@ -47,4 +52,9 @@ class Build(BaseCommand):
         action.execute(outline=options.outline, tail=options.tail)
 
     def get_context_kwargs(self, options, **kwargs):
-        return {'stack_names': options.stacks, 'force_stacks': options.force}
+        return {
+            'stack_names': options.stacks,
+            'force_stacks': options.force,
+            'skip_hooks': options.skip_hook,
+            'run_hook': options.run_hook,
+        }